unsiged int for binary

+++ _printf.c
@@ -15,7 +15,7 @@ int _printf(const char *format, ...)
 	va_list ap;
 	char *temp;
 	int i = 0, num, count = 0;
-	unsigned long int num2;
+	unsigned int num2;
 
 	if (format == NULL)
 		return (-1);
@@ -48,7 +48,7 @@ int _printf(const char *format, ...)
 					count += print_number(num);
 				break;
 			case 'b':
-				num2 = va_arg(ap, unsigned long int);
+				num2 = va_arg(ap, unsigned int);
 				count += int_to_bin(num2);
 				break;
 			default:
